+++ crates/matrix-sdk-base/src/room/room_info.rs
@@ -1412,27 +1412,27 @@ pub struct RoomInfoNotableUpdate {
 bitflags! {
     /// The reason why a [`RoomInfoNotableUpdate`] is emitted.
     #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-    pub struct RoomInfoNotableUpdateReasons: u8 {
+    pub struct RoomInfoNotableUpdateReasons: u16 {
         /// The recency stamp of the `Room` has changed.
-        const RECENCY_STAMP = 0b0000_0001;
+        const RECENCY_STAMP = 0b0000_0000_0000_0001;
 
         /// The latest event of the `Room` has changed.
-        const LATEST_EVENT = 0b0000_0010;
+        const LATEST_EVENT = 0b0000_0000_0000_0010;
 
         /// A read receipt has changed.
-        const READ_RECEIPT = 0b0000_0100;
+        const READ_RECEIPT = 0b0000_0000_0000_0100;
 
         /// The user-controlled unread marker value has changed.
-        const UNREAD_MARKER = 0b0000_1000;
+        const UNREAD_MARKER = 0b0000_0000_0000_1000;
 
         /// A membership change happened for the current user.
-        const MEMBERSHIP = 0b0001_0000;
+        const MEMBERSHIP = 0b0000_0000_0001_0000;
 
         /// The display name has changed.
-        const DISPLAY_NAME = 0b0010_0000;
+        const DISPLAY_NAME = 0b0000_0000_0010_0000;
 
         /// The active service members have changed.
-        const ACTIVE_SERVICE_MEMBERS = 0b0100_0000;
+        const ACTIVE_SERVICE_MEMBERS = 0b0000_0000_0100_0000;
 
         /// This is a temporary hack.
         ///
@@ -1444,7 +1444,10 @@ bitflags! {
         /// identified, we are likely to miss particular updates, and it can feel broken.
         /// Ultimately, we want to clearly identify all the notable update reasons, and
         /// remove this one.
-        const NONE = 0b1000_0000;
+        const NONE = 0b0000_0000_1000_0000;
+
+        /// The user's `m.fully_read` marker has changed.
+        const FULLY_READ = 0b0000_0001_0000_0000;
     }
 }