+++ crates/matrix-sdk-base/CHANGELOG.md
@@ -14,6 +14,9 @@ All notable changes to this project will be documented in this file.
   This is needed for calculating display names of `SpaceRoom`s with service members. 
   ([#6561](https://github.com/matrix-org/matrix-rust-sdk/pull/6561))
 
+- [**breaking**] `RoomInfoNotableUpdateReasons` is now a `u16` to include a `FULLY_READ` flag to notify on changes of the `m.fully_read` marker.
+  ([#6569](https://github.com/matrix-org/matrix-rust-sdk/pull/6569))
+
 ## [0.17.0] - 2026-05-08
 
 ### Bug Fixes